Reduce tearing effect on the h747i-disco
There seems to be an issue with the reload during vertical blank not happening at a suitable time, when the
screen is operated in landscape mode,
as exhibited in https://community.st.com/t5/stm32-mcus-products/stm32h747i-disco-tearing-in-double-buffering/td-p/647508 and even the ST provided LCD_DSI_VideoMode_DoubleBuffering example.

One workaround that seems to work well is to drive the display in portrait mode instead, so let's do that.